Moving health and opmode related calls to health utils, minor linting
