+++ packages/react-components/lib/doors/door-card.tsx
@@ -1,7 +1,7 @@
 import { Card, CardContent, CardProps, Grid, SxProps, Typography, useTheme } from '@mui/material';
 import React from 'react';
 import { DoorMode } from 'rmf-models';
-import { doorModeToString, doorTypeToString } from './utils';
+import { doorModeToString, doorTypeToString } from './door-utils';
 
 export interface DoorCardProps extends CardProps {
   name: string;
