+++ packages/react-components/lib/doors/index.ts
@@ -1,4 +1,4 @@
 export * from './door-card';
 export * from './door-controls';
-export * from './utils';
+export * from './door-utils';
 export * from './door-table-datagrid';
