+++ packages/react-components/lib/lifts/lift-table-datagrid.tsx
@@ -5,12 +5,8 @@ import { Lift } from 'api-client';
 import ArrowDownwardIcon from '@mui/icons-material/ArrowDownward';
 import ArrowUpwardIcon from '@mui/icons-material/ArrowUpward';
 import { LiftState as LiftStateModel } from 'rmf-models';
-import {
-  HealthStatus,
-  doorStateToString,
-  getHealthStatusDescription,
-  motionStateToString,
-} from './lift-utils';
+import { doorStateToString, motionStateToString } from './lift-utils';
+import { HealthStatus, healthStatusToOpMode } from '../utils';
 import { LiftControls } from './lift-controls';
 
 export interface LiftTableData {
@@ -74,7 +70,7 @@ export function LiftDataGridTable({ lifts }: LiftDataGridTableProps): JSX.Elemen
             fontSize: 14,
           }}
         >
-          {getHealthStatusDescription(params.row.opMode)}
+          {healthStatusToOpMode(params.row.opMode)}
         </Typography>
       </Box>
     );
