+++ packages/react-components/lib/utils/index.ts
@@ -1,3 +1,4 @@
 export * from './geometry';
-export * from './misc';
+export * from './health';
 export * from './item-table';
+export * from './misc';
